split most streamers into own files
